requirements.txt
chromadb>=0.4.22
openai>=1.12.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0
rich>=13.7.0